requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==6.2.1  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
whitenoise==4.1.4  # https://github.com/evansd/whitenoise
redis==3.3.11  # https://github.com/antirez/redis

# Django
# ------------------------------------------------------------------------------
django==2.2.10  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.40.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.8.1  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis
django-filter==2.0.0  # https://github.com/carltongibson/django-filter

# Django REST Framework
djangorestframework==3.10.3  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
django-rest-swagger==2.2.0
drf-yasg==1.17.0 # https://github.com/axnsan12/drf-yasg
